Do not inline stream builders for JS because of ReferenceError in tests
