+++ reaktive/src/jsMain/kotlin/com/badoo/reaktive/maybe/MaybeByEmitter.kt
@@ -1,0 +1,46 @@
+package com.badoo.reaktive.maybe
+
+import com.badoo.reaktive.base.tryCatch
+import com.badoo.reaktive.disposable.Disposable
+import com.badoo.reaktive.disposable.DisposableWrapper
+
+actual fun <T> maybe(onSubscribe: (emitter: MaybeEmitter<T>) -> Unit): Maybe<T> =
+    maybeUnsafe { observer ->
+        val emitter =
+            object : DisposableWrapper(), MaybeEmitter<T> {
+                override fun setDisposable(disposable: Disposable?) {
+                    set(disposable)
+                }
+
+                override fun onSuccess(value: T) {
+                    doIfNotDisposedAndDispose {
+                        observer.onSuccess(value)
+                    }
+                }
+
+                override fun onComplete() {
+                    doIfNotDisposedAndDispose(observer::onComplete)
+                }
+
+                override fun onError(error: Throwable) {
+                    doIfNotDisposedAndDispose {
+                        observer.onError(error)
+                    }
+                }
+
+                private inline fun doIfNotDisposedAndDispose(block: () -> Unit) {
+                    if (!isDisposed) {
+                        val disposable: Disposable? = replace(null)
+                        try {
+                            dispose()
+                            block()
+                        } finally {
+                            disposable?.dispose()
+                        }
+                    }
+                }
+            }
+
+        observer.onSubscribe(emitter)
+        emitter.tryCatch { onSubscribe(emitter) }
+    }
